+++ src/include/definitions/colors.c
@@ -1,29 +1,5 @@
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
 #include <stdbool.h>
-#include "color_definitions.h"
-#include "lowfetch_base.h"
-
-char *file_read(const char *filename, size_t size)
-{
-    FILE *file = fopen(filename, "r");
-    if (!file)
-    {
-        fprintf(stderr, "error: '%s' is inaccessible or does not exist\n", filename);
-        return NULL;
-    }
-
-    char *output;
-    output = malloc((size+1)*sizeof(*output)); 
-
-    fread(output, size, 1, file); // store the file into output
-    output[strlen(output) - 1] = 0; // trim the trailing newline
-
-    fclose(file); // close the file since it's already been stored
-
-    return output;
-}
+#include "colors.h"
 
 char *get_ansi_color_from(char color_char, bool bold)
 {
@@ -31,8 +7,6 @@ char *get_ansi_color_from(char color_char, bool bold)
     {
         switch (color_char)
         {
-            case CHAR_WHITE:
-                return ANSI_COLOR_RESET;
             case CHAR_RED:
                 return ANSI_COLOR_RED;
             case CHAR_GREEN:
@@ -53,8 +27,6 @@ char *get_ansi_color_from(char color_char, bool bold)
     {
         switch (color_char)
         {
-            case CHAR_WHITE:
-                return ANSI_COLOR_RESET_BOLD;
             case CHAR_RED:
                 return ANSI_COLOR_RED_BOLD;
             case CHAR_GREEN:
@@ -71,5 +43,5 @@ char *get_ansi_color_from(char color_char, bool bold)
                 return ANSI_COLOR_RESET_BOLD;
         }
     }
+    return ANSI_COLOR_RESET;
 }
-
