+++ src/include/definitions/colors.h
@@ -1,5 +1,6 @@
-#ifndef COLOR_DEFINITIONS_H_
-#define COLOR_DEFINITIONS_H_
+#ifndef COLORS_H_
+#define COLORS_H_
+#include <stdbool.h>
 
 // color chars
 #define CHAR_WHITE   'w'
@@ -10,7 +11,7 @@
 #define CHAR_MAGENTA 'm'
 #define CHAR_CYAN    'c'
 
-// colors
+// color codes
 #define ANSI_COLOR_RED     "\x1b[31m"
 #define ANSI_COLOR_GREEN   "\x1b[32m"
 #define ANSI_COLOR_YELLOW  "\x1b[33m"
@@ -19,7 +20,6 @@
 #define ANSI_COLOR_CYAN    "\x1b[36m"
 #define ANSI_COLOR_RESET   "\x1b[0m"
 
-// bold colors
 #define ANSI_COLOR_RED_BOLD     "\x1b[31;1m"
 #define ANSI_COLOR_GREEN_BOLD   "\x1b[32;1m"
 #define ANSI_COLOR_YELLOW_BOLD  "\x1b[33;1m"
@@ -27,5 +27,7 @@
 #define ANSI_COLOR_MAGENTA_BOLD "\x1b[35;1m"
 #define ANSI_COLOR_CYAN_BOLD    "\x1b[36;1m"
 #define ANSI_COLOR_RESET_BOLD   "\x1b[0;1m"
+
+char *get_ansi_color_from(char color_char, bool bold);
 
 #endif
