+++ src/include/definitions/sizes.h
@@ -1,0 +1,12 @@
+#ifndef SIZES_H_
+#define SIZES_H_
+
+#define ASCII_FILESIZE      4096
+#define KERNEL_SIZE         512
+#define FILENAME_SIZE       256
+#define DISTRO_ID_SIZE      256
+#define UPTIME_SIZE         256
+#define ORDER_FILESIZE      128
+#define PACKAGE_AMOUNT_SIZE 32
+
+#endif
