+++ src/include/distro/distro.h
@@ -1,0 +1,7 @@
+#ifndef DISTRO_H_
+#define DISTRO_H_
+#include <stdlib.h>
+
+char *get_distro_id(size_t size);
+
+#endif
