+++ src/include/general/file_read.h
@@ -1,0 +1,7 @@
+#ifndef FILE_READ_H_
+#define FILE_READ_H_
+#include <stdlib.h>
+
+char *file_read(const char *filename, size_t size);
+
+#endif
