+++ src/include/shell/shell.h
@@ -1,0 +1,6 @@
+#ifndef SHELL_H_
+#define SHELL_H_
+
+char *get_shell();
+
+#endif
