--- src/include/size_definitions.h
@@ -1,12 +1,0 @@
-#ifndef SIZE_DEFINITIONS_H_
-#define SIZE_DEFINITIONS_H_
-
-#define ASCII_FILESIZE 4096
-#define KERNEL_VERSION_SIZE 512
-#define FILENAME_SIZE 256
-#define DISTRO_ID_SIZE 256
-#define UPTIME_SIZE 256
-#define ORDER_FILESIZE 128
-#define PACKAGE_AMOUNT_SIZE 32
-
-#endif
